ConverterContextTesting was ConverterLikeTesting & ConverterLikeTesting was ConvertLikeTesting
- Previous commit renamed XXXTesting classes badly, this commit fixes that.
