+++ src/main/java/walkingkooka/convert/ConverterLikeTesting.java
@@ -16,26 +16,84 @@
  */
 package walkingkooka.convert;
 
-import walkingkooka.datetime.DateTimeContextTesting2;
-import walkingkooka.math.DecimalNumberContextTesting2;
+import org.opentest4j.AssertionFailedError;
+import walkingkooka.Either;
+import walkingkooka.test.Testing;
+import walkingkooka.text.CharSequences;
 
 /**
- * Mixing testing interface for {@link ConverterContext}
+ * Mixing testing interface for {@link ConverterLike}
  */
-public interface ConverterLikeTesting<C extends ConverterContext> extends ConvertLikeTesting<C>,
-    DateTimeContextTesting2<C>,
-    DecimalNumberContextTesting2<C> {
+public interface ConverterLikeTesting<C extends ConverterLike> extends Testing {
+
+    default <T> T convertAndCheck(final Object value,
+                                  final Class<T> target,
+                                  final T expected) {
+        return this.convertAndCheck(this.createConverterLike(), value, target, expected);
+    }
+
+    default <T> T convertAndCheck(final ConverterLike context,
+                                  final Object value,
+                                  final Class<T> target,
+                                  final T expected) {
+        this.checkEquals(
+            true,
+            context.canConvert(value, target),
+            () -> context + " can convert(" + CharSequences.quoteIfChars(value) + "(" + value.getClass().getName() + ")," + target.getName() + ")"
+        );
+
+        final Either<T, String> result = context.convert(value, target);
+        if (result.isRight()) {
+            throw new AssertionFailedError(result.rightValue());
+        }
+
+        final T convertedValue = result.leftValue();
+        this.checkEquals(
+            expected,
+            convertedValue,
+            () -> "Failed to convert " + CharSequences.quoteIfChars(value) + " (" + value.getClass().getName() + ")= to " + target.getName()
+        );
+        return convertedValue;
+    }
+
+    default void convertFails(final Object value,
+                              final Class<?> type) {
+        this.convertFails(this.createConverterLike(), value, type);
+    }
+
+    default void convertFails(final ConverterLike context,
+                              final Object value,
+                              final Class<?> type) {
+        final Either<?, String> result = context.convert(value, type);
+        result.mapLeft(v -> {
+            throw new AssertionFailedError("Expected failure converting " + CharSequences.quoteIfChars(value) + " to " + type.getName() + " but got " + CharSequences.quoteIfChars(v));
+        });
+    }
+
+    default <T> T convertOrFailAndCheck(final Object value,
+                                        final Class<T> target,
+                                        final T expected) {
+        return this.convertOrFailAndCheck(this.createConverterLike(),
+            value,
+            target,
+            expected);
+    }
+
+    default <T> T convertOrFailAndCheck(final ConverterLike can,
+                                        final Object value,
+                                        final Class<T> target,
+                                        final T expected) {
+        final T convertedValue = can.convertOrFail(value, target);
+        this.checkEquals(
+            expected,
+            convertedValue,
+            () -> "Failed to convertOrFail " + CharSequences.quoteIfChars(value) + " (" + value.getClass().getName() + ")= to " + target.getName()
+        );
+        return convertedValue;
+    }
 
     /**
-     * Delegates to {@link #createContext()}.
+     * Factory that creates a {@link ConverterLike}.
      */
-    @Override
-    default C createConverterLike() {
-        return this.createContext();
-    }
-
-    @Override
-    default String typeNameSuffix() {
-        return ConverterContext.class.getSimpleName();
-    }
+    C createConverterLike();
 }
